requirements.txt
asgiref
cffi
cryptography
django==3.2
django-rest-knox
djangorestframework==3.13.1
django-cors-headers
pycparser
pytz
sqlparse
typing_extensions